backend/requirements.txt
# Django & DRF
Django==5.0.1
djangorestframework==3.14.0
django-filter==23.5

# Authentication
djangorestframework-simplejwt==5.3.1

# CORS
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Image handling
Pillow==10.2.0

# Environment variables
python-decouple==3.8

# Production server
gunicorn==21.2.0

# Gemini AI Integration
google-genai>=1.52.0

# Storage (optional - pour AWS S3 si besoin)
# django-storages==1.14.2
# boto3==1.34.22